blabla

+++ putback.go
@@ -4,17 +4,21 @@ import (
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
 	"log"
-	"strings"
 	"vidego/pkg/datatype"
 	"vidego/pkg/utils"
 )
 
-var sql_request_putback = `select path, name, duration, size
+var sql_request_putback = `select *
 							from videogo.video
-							where path like '%dedup%';`
+								where duration in (select duration from videogo.video group by duration having count(1) > 1)`
 
+/*
+`select path, name, duration, size
+from videogo.video
+where path like '%dedup%';`
+*/
 func main() {
-	dsn := "host=db.mend.ovh user=fabien password=xxoca306 dbname=videogo port=5434 sslmode=disable"
+	dsn := "host=db.mend.ovh user=fabien password=xxoca306 dbname=fabien port=5434 sslmode=disable"
 	db, err := gorm.Open(postgres.Open(dsn), &gorm.Config{})
 
 	if err != nil {
@@ -29,14 +33,13 @@ func main() {
 	for _, dedup := range dedups {
 		log.Printf("putback %s\n", dedup.Name)
 
-		source := dedup.Path + "/" + dedup.Name
+		source := dedup.Path + "/dedup/" + dedup.Name
+		dest := dedup.Path + "/" + dedup.Name
 
-		dest := strings.ReplaceAll(dedup.Path, "/dedup", "") + "/" + dedup.Name
+		//dest := strings.ReplaceAll(dedup.Path, "/dedup", "") + "/" + dedup.Name
 		result := utils.MoveFile(source, dest)
-		if result {
-			// update db
-			db.Save(&dedup).Update("path", dest)
-		} else {
+		existsInDest := utils.Exists(dest)
+		if !result && !existsInDest {
 			// remove from db
 			db.Delete(&dedup)
 		}
